Fix package names
