+++ bundles/org.openhab.binding.logreader/pom.xml
@@ -18,7 +18,6 @@
       <plugin>
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>build-helper-maven-plugin</artifactId>
-        <version>3.0.0</version>
         <executions>
           <execution>
             <phase>generate-sources</phase>
