+++ bundles/org.openhab.binding.logreader/src/3rdparty/java/org/openhab/binding/logreader/internal/commonsio/Tailer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openhab.binding.logreader.internal.thirdparty.commonsio;
+package org.openhab.binding.logreader.internal.commonsio;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
