+++ bundles/org.openhab.binding.logreader/src/main/java/org/openhab/binding/logreader/internal/filereader/FileTailer.java
@@ -16,11 +16,11 @@ import java.io.File;
 import java.util.concurrent.ScheduledExecutorService;
 
 import org.eclipse.jdt.annotation.Nullable;
+import org.openhab.binding.logreader.internal.commonsio.Tailer;
+import org.openhab.binding.logreader.internal.commonsio.TailerListener;
+import org.openhab.binding.logreader.internal.commonsio.TailerListenerAdapter;
 import org.openhab.binding.logreader.internal.filereader.api.FileReaderException;
 import org.openhab.binding.logreader.internal.filereader.api.LogFileReader;
-import org.openhab.binding.logreader.internal.thirdparty.commonsio.Tailer;
-import org.openhab.binding.logreader.internal.thirdparty.commonsio.TailerListener;
-import org.openhab.binding.logreader.internal.thirdparty.commonsio.TailerListenerAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
